+++ paypal/standard/ipn/tests/test_ipn.py
@@ -22,12 +22,14 @@ from paypal.standard.ipn.signals import (payment_was_successful,
 # Parameters are all bytestrings, so we can construct a bytestring
 # request the same way that Paypal does.
 
+TEST_RECEIVER_EMAIL = b"seller@paypalsandbox.com"
+
 CHARSET = "windows-1252"
 IPN_POST_PARAMS = {
     "protection_eligibility": b"Ineligible",
     "last_name": b"User",
     "txn_id": b"51403485VH153354B",
-    "receiver_email": b(settings.PAYPAL_RECEIVER_EMAIL),
+    "receiver_email": TEST_RECEIVER_EMAIL,
     "payment_status": b"Completed",
     "payment_gross": b"10.00",
     "tax": b"0.00",
@@ -207,11 +209,6 @@ class IPNTest(IPNTestBase):
 
         self.assertGotSignal(payment_was_reversed, False, params, deprecated=True)
 
-    def test_incorrect_receiver_email(self):
-        update = {"receiver_email": "incorrect_email@someotherbusiness.com"}
-        flag_info = "Invalid receiver_email. (incorrect_email@someotherbusiness.com)"
-        self.assertFlagged(update, flag_info)
-
     def test_invalid_payment_status(self):
         update = {"payment_status": "Failure"}
         flag_info = u"Invalid payment_status. (Failure)"
@@ -381,8 +378,7 @@ class IPNPostbackTest(IPNTestBase):
         self.assertFlagged({}, u'Invalid postback. (INVALID)')
 
 
-@override_settings(ROOT_URLCONF='paypal.standard.ipn.tests.test_urls',
-                   PAYPAL_RECEIVER_EMAIL='seller@paypalsandbox.com')
+@override_settings(ROOT_URLCONF='paypal.standard.ipn.tests.test_urls')
 class IPNSimulatorTests(TestCase):
 
     # Some requests, as sent by the simulator.